Moved install button text to Resources ProgressTypes -> Progress
